// src/models/MyBooks/BookListSortFilter.ts
export enum BookListSortNames {
  RECENT = 'Recent',
  DATE_ADDED = 'Date Added',
  DATE_FINISHED = 'Date Finished',
  DATE_PUBLISHED = 'Date Published',
  TITLE = 'Title',
  AUTHOR = 'Author',
}

export enum BookListFormatNames {
  ALL = 'All',
  EBOOKS = 'ebook',
  AUDIOBOOKS = 'audiobook'
}

export enum BookListBookStatusNames {
  TO_BE_READ = 'To Be Read',
  READING = 'Reading',
  FINISHED = 'Finished',
}